+++ server/src/main/java/org/allaymc/server/player/AllayPlayer.java
@@ -198,6 +198,9 @@ public class AllayPlayer implements Player {
     protected boolean clientCacheEnabled;
     protected boolean shouldSendCommands;
     @Getter
+    @Setter
+    protected boolean containerClosedByClient;
+    @Getter
     protected Speed speed, flySpeed, verticalFlySpeed;
 
     // Container
@@ -1663,6 +1666,13 @@ public class AllayPlayer implements Player {
         var packet = new ContainerClosePacket();
         packet.setId(assignedId);
         packet.setType(ContainerNetworkInfo.getInfo(container.getContainerType()).toNetworkType());
+        if (!this.containerClosedByClient) {
+            // Field `serverInitiated` determines whether the server force-closed the container. If this value is
+            // not set correctly, the client may ignore the packet and respond with a `PacketViolationWarningPacket`.
+            packet.setServerInitiated(true);
+        } else {
+            this.containerClosedByClient = false;
+        }
         sendPacket(packet);
     }